backend/app/moderation.py
from PIL import Image
import requests
from io import BytesIO
from fastapi import HTTPException

def moderate_image(image_url: str) -> dict:
    """
    Placeholder for image moderation logic.
    Returns a mock content-safety report based on image metadata.
    """
    try:
        response = requests.get(image_url)
        response.raise_for_status()
        img = Image.open(BytesIO(response.content))
        
        # Mock moderation logic: Check image size and format
        is_safe = img.size[0] < 2000 and img.size[1] < 2000 and img.format in ["JPEG", "PNG"]
        categories = {
            "violence": 0.0 if is_safe else 0.8,
            "nudity": 0.0 if is_safe else 0.5,
            "hate_symbols": 0.0 if is_safe else 0.3,
            "self_harm": 0.0 if is_safe else 0.2,
            "extremist": 0.0 if is_safe else 0.1
        }
        return {"is_safe": is_safe, "categories": categories}
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Image processing failed: {str(e)}")