backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
pillow==10.4.0
pydantic==2.9.2